requirements.txt
requests
PySocks
pytz